# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CampusForge API
  version: 0.1.0
servers:
  - url: http://localhost:4000/v1
  - url: http://localhost:3000/api
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: campusforge-api
                  version:
                    type: string
                    example: "0.1.0"
  /applications:
    post:
      summary: Submit an invitation application
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplication"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid request
        "409":
          description: User already has an application
    get:
      summary: List all applications (admin)
      responses:
        "200":
          description: Application list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Application"
                  total:
                    type: integer
  /applications/me:
    get:
      summary: Get own application
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: No application found
  /applications/{id}:
    get:
      summary: Get application by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Application not found
  /applications/{id}/review:
    patch:
      summary: Review an application (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApplication"
      responses:
        "200":
          description: Application reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Application not found
        "409":
          description: Invalid review status transition
  /github/issues:
    get:
      summary: List open GitHub issues for the configured organization
      responses:
        "200":
          description: GitHub issue list
          content:
            application/json:
              schema:
                type: object
                properties:
                  org:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/GitHubIssue"
        "401":
          description: Login required
        "502":
          description: GitHub request failed
  /github/issues/claim:
    post:
      summary: Claim a GitHub issue and assign it remotely when configured
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimGitHubIssue"
      responses:
        "200":
          description: Issue assigned on GitHub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimGitHubIssueResult"
        "202":
          description: Local claim accepted, remote assignment skipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimGitHubIssueResult"
        "400":
          description: Invalid claim request
        "401":
          description: Login required
        "502":
          description: GitHub assignment failed
    delete:
      summary: Cancel a local GitHub issue claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimGitHubIssue"
      responses:
        "200":
          description: Claim cancelled
        "400":
          description: Invalid claim request
        "401":
          description: Login required
        "403":
          description: Only the claimant or an admin can cancel the claim
  /github/profile:
    get:
      summary: Get cached GitHub profile contribution stats
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cached or freshly fetched profile statistics
        "400":
          description: Missing username
        "401":
          description: Login required
        "502":
          description: GitHub profile request failed
  /github/connection:
    get:
      summary: Get the current user's verified GitHub connection
      responses:
        "200":
          description: GitHub connection, or null when not connected
        "401":
          description: Login required
    delete:
      summary: Disconnect the current user's GitHub account
      responses:
        "200":
          description: GitHub connection removed
        "401":
          description: Login required
  /github/oauth/start:
    get:
      summary: Start GitHub OAuth account connection
      responses:
        "307":
          description: Redirect to GitHub authorization
        "401":
          description: Login required
  /github/oauth/callback:
    get:
      summary: Complete GitHub OAuth account connection
      responses:
        "307":
          description: Redirect back to profile after connection
components:
  schemas:
    CreateApplication:
      type: object
      required: [schoolEmail, studentId, department, reason]
      properties:
        schoolEmail:
          type: string
          format: email
        studentId:
          type: string
        department:
          type: string
        reason:
          type: string
          minLength: 10
          maxLength: 1000
    ReviewApplication:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED, NEEDS_INFO]
        reviewReason:
          type: string
          maxLength: 1000
    Application:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        schoolEmail:
          type: string
        studentId:
          type: string
        department:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [DRAFT, SUBMITTED, VERIFYING, UNDER_REVIEW, APPROVED, REJECTED, NEEDS_INFO]
        reviewerId:
          type: string
          nullable: true
        reviewReason:
          type: string
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GitHubIssue:
      type: object
      properties:
        id:
          type: string
        owner:
          type: string
        repo:
          type: string
        number:
          type: integer
        title:
          type: string
        url:
          type: string
        state:
          type: string
        comments:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        labels:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              color:
                type: string
        assignees:
          type: array
          items:
            type: string
        claimedBy:
          type: string
          nullable: true
        claimedAt:
          type: string
          format: date-time
          nullable: true
        remoteAssigned:
          type: boolean
        author:
          type: string
    ClaimGitHubIssue:
      type: object
      required: [owner, repo, number, assignee]
      properties:
        owner:
          type: string
        repo:
          type: string
        number:
          type: integer
        title:
          type: string
        url:
          type: string
        assignee:
          type: string
    ClaimGitHubIssueResult:
      type: object
      properties:
        remoteAssigned:
          type: boolean
        claimedBy:
          type: string
        claimedAt:
          type: string
          format: date-time
        reason:
          type: string